// frontend/src/types/api.type.ts
export type TaskType = {
  _id: string;
  title: string;
  description?: string;
  project?: {
    _id: string;
    emoji: string;
    name: string;
  };
  priority: TaskPriorityEnumType;
  status: TaskStatusEnumType;
  assignedTo: {
    _id: string;
    name: string;
    profilePicture: string | null;
  } | null;
  createdBy?: {
    _id: string;
    name: string;
    profilePicture: string | null;
  };
  dueDate: string;
  taskCode: string;
  createdAt?: string;
  updatedAt?: string;
};